[site] Added Footer with required functionality
Now every page rendered has a Footer island. It contains Copyright
notice and links to PP and ToS as required by the subject. Because those
are links also implemented routes that return those pages and the
contents as separate components.

The content of these has to still be updated separately but knowing that
requires almost the entire project to be completed so this just adds the
necessary functionality.